day end thursday

+++ backend/products/models.py
@@ -18,7 +18,7 @@ class Product(models.Model):
     description = models.TextField(blank=True)
     price = models.IntegerField()
     bedrooms = models.IntegerField()
-    bathrooms = models.DecimalField(max_digits=2, decimal_places=1)
+    bathrooms = models.IntegerField()
     sqft = models.IntegerField()
     photo_main = models.ImageField(upload_to='photos/%Y/%m/%d')
     photo = models.ManyToManyField(ProductImages,  blank=True, null=True,)
